engine/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-socketio[asgi]==5.12.1
pydantic==2.10.6
